0.0.1v_최초 버전 배포

+++ Motor.h
@@ -48,7 +48,7 @@ public:
 
     struct MotorSetting {
         uint8_t dir;
-        uint16_t angle;
+        float angle;
         uint16_t initPosi;
         uint16_t reducer_ratio;
     };
